Shift fixups

+++ src/shifting_algo/s_algo.h
@@ -145,7 +145,7 @@ public:
         virtual uint16_t high_fill_pressure() = 0;
         virtual bool is_release_shift() = 0;
         uint16_t threshold_rpm = 0;
-        uint16_t spc_p_offset = 0;
+        float spc_p_offset = 0;
 };
 
 // Helper functions
